Clean up packet inspector a bit
