+++ Cargo.toml
@@ -126,7 +126,6 @@ directories = "5.0.0"
 eframe = { version = "0.22.0", default-features = false }
 egui = "0.22.0"
 egui_dock = "0.6"
-enum-map = "2.5.0"
 flate2 = "1.0.24"
 flume = "0.11.0"
 fs_extra = "1.2.0"
