+++ crates/valence_protocol/src/lib.rs
@@ -78,6 +78,7 @@ pub use ident::ident;
 pub use item::{ItemKind, ItemStack};
 pub use packets::play::particle_s2c::Particle;
 pub use raw::RawBytes;
+use serde::{Deserialize, Serialize};
 pub use sound::Sound;
 pub use text::Text;
 pub use valence_generated::{block, packet_id};
@@ -295,7 +296,7 @@ pub trait Packet: std::fmt::Debug {
 }
 
 /// The side a packet is intended for.
-#[derive(Copy, Clone, PartialEq, Eq)]
+#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug, Serialize, Deserialize)]
 pub enum PacketSide {
     /// Server -> Client
     Clientbound,
@@ -304,7 +305,7 @@ pub enum PacketSide {
 }
 
 /// The statein  which a packet is used.
-#[derive(Copy, Clone, PartialEq, Eq)]
+#[derive(Copy, Clone, PartialEq, Eq, Hash, Debug, Serialize, Deserialize)]
 pub enum PacketState {
     Handshaking,
     Status,
