+++ tools/packet_inspector/src/app/text_viewer.rs
@@ -1,7 +1,7 @@
 use super::{SharedState, Tab, View};
 
 mod utils {
-    use packet_inspector::{Packet as ProxyPacket, PacketSide, PacketState};
+    use packet_inspector::Packet as ProxyPacket;
     use valence_protocol::packets::handshaking::*;
     use valence_protocol::packets::login::*;
     use valence_protocol::packets::play::*;
