+++ tools/packet_inspector/src/packet_registry.rs
@@ -4,7 +4,7 @@ use std::sync::RwLock;
 use bytes::Bytes;
 use time::OffsetDateTime;
 use valence_protocol::decode::PacketFrame;
-use valence_protocol::CompressionThreshold;
+use valence_protocol::{CompressionThreshold, PacketSide, PacketState};
 
 pub struct PacketRegistry {
     packets: RwLock<Vec<Packet>>,
@@ -88,18 +88,17 @@ impl Default for PacketRegistry {
     }
 }
 
-#[derive(Clone, Debug, Eq)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[derive(Clone, Debug, Eq, serde::Serialize, serde::Deserialize)]
 pub struct Packet {
     pub side: PacketSide,
     pub state: PacketState,
     pub id: i32,
-    #[cfg_attr(feature = "serde", serde[skip])]
+    #[serde(skip)]
     pub timestamp: Option<OffsetDateTime>,
-    #[cfg_attr(feature = "serde", serde[skip])]
+    #[serde(skip)]
     pub name: &'static str,
     /// Uncompressed packet data
-    #[cfg_attr(feature = "serde", serde[skip])]
+    #[serde(skip)]
     pub data: Option<Bytes>,
 }
 
@@ -118,20 +117,4 @@ impl Hash for Packet {
         self.side.hash(state);
         self.state.hash(state);
     }
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub enum PacketState {
-    Handshaking,
-    Status,
-    Login,
-    Play,
-}
-
-#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-pub enum PacketSide {
-    Clientbound,
-    Serverbound,
 }
